Move `BetaFeaturesConfigurationViewModel` to its own file.
